Rtp 92 le joueur peut jouer a un r type en ligne

+++ src/SharedNetwork/Utils/UserMessage.hpp
@@ -5,24 +5,24 @@
 #pragma once
 
 #include <memory>
-#include "IMessage.hpp"
+#include "Message.hpp"
 
 namespace Network {
 
     struct UserMessage {
-        size_t id;
+        unsigned int id;
         std::shared_ptr<IMessage> message;
-        UserMessage(size_t id, std::shared_ptr<IMessage> message) : id(id), message(message) {};
+        UserMessage(unsigned int id, std::shared_ptr<IMessage> message) : id(id), message(message) {};
     };
     struct UsersMessage {
-        std::vector<size_t> ids;
+        std::vector<unsigned int> ids;
         std::shared_ptr<IMessage> message;
-        UsersMessage(std::vector<size_t> ids, std::shared_ptr<IMessage> message) : ids(ids), message(message) {};
+        UsersMessage(std::vector<unsigned int> ids, std::shared_ptr<IMessage> message) : ids(ids), message(message) {};
     };
     struct NotUserMessage {
-        size_t id;
+        unsigned int id;
         std::shared_ptr<IMessage> message;
-        NotUserMessage(size_t id, std::shared_ptr<IMessage> message) : id(id), message(message) {};
+        NotUserMessage(unsigned int id, std::shared_ptr<IMessage> message) : id(id), message(message) {};
     };
     struct AllUsersMessage {
         std::shared_ptr<IMessage> message;
